+++ NEWS.md
@@ -46,6 +46,7 @@ Many bugs have been fixed, and numerous features that were only partially implem
 * Improved input device configuration, and default keyboard mappings are now applied if the keyboard has not been configured by the user
 * GUI-configurable option to sort favorite games above non-favorite games (favorites marked with stars)
 * GUI-configurable option to sort folders on top of the gamelists
+* Added a gamelist info text field displaying the game count, any applied filters as well as an icon if a folder has been entered
 * Expanded the metadata for folders and made it possible to mark them as favorites
 * Added new component GuiComplexTextEditPopup to handle changes to configuration file entries and similar
 * Speed improvements and optimizations, the application now starts faster and feels more responsive